Better Detect MetaMask (#16)
The constructor on the MetaMask object for me is returning `g`, which indicates the extension is probably being compiled with minification. In that case, we were returning `null` from `providerType` which incorrectly assumed we should then auto-connect. As such, we now always return some variety of MetaMask-type even if we can't positively identify the constructor for mobile versus web.