Remove Node-Sass Chokidar
This patch removes `node-sass-chokidar` which should be removed, if possible, from the planet. Node Sass is the node-gyp compiled version of libsass which has a ton of build issues, but is slow in and of itself even if you get it working. Google created dart-sass (now called just `sass` in npm) that is pure JavaScript, fast, and doesn't require C-builds. On top of that, this project doesn't even compile sass (look in the app, I can't find where it's calling into sass compilation- the dep was being used by code that referenced this library, but shouldn't be in this library itself). Finally, chokidar is weird and there are better ways to listen for filesystem events in more modern versions of macOS. As such: Be Gone, Nodesass Chokidar.